requirements.txt
asgiref==3.6.0
Babel==2.12.1
backports.zoneinfo==0.2.1
boto3==1.26.133
botocore==1.29.133
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-money==3.1.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
py-moneyed==2.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
